requirements.txt
numpy
python-dotenv
telebot